% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_functions.R
\name{call.tcga.cnas}
\alias{call.tcga.cnas}
\title{Call copy numbers from TCGA level 3 data}
\usage{
call.tcga.cnas(cnv.data, colour.code = FALSE)
}
\arguments{
\item{cnv.data}{Data frame of TCGA Level 3 data}

\item{colour.code}{Logical indicating if colours should be returned instead of -1, 0, 1 coding}
}
\value{
vector of CNA calls
}
\description{
Call copy numbers from TCGA level 3 data
}
