% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{censoring.km.estimate}
\alias{censoring.km.estimate}
\title{Get Kaplan-Meier estimate of censoring distribution at time t.}
\usage{
censoring.km.estimate(t, time, status, censoring.status = 0)
}
\arguments{
\item{t}{time}

\item{time}{vector of event times}

\item{status}{vector of event status}

\item{censoring.status}{status corresponding to censored observations, defaults to zero.}
}
\value{
km.estimate
}
\description{
Get Kaplan-Meier estimate of censoring distribution at time t.
}

