% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_functions.R
\name{colour.code.cnas}
\alias{colour.code.cnas}
\title{Colour code CNAs}
\usage{
colour.code.cnas(cna.calls, loss.col = "blue", gain.col = "red",
  neutral.col = "white")
}
\arguments{
\item{cna.calls}{vector of -1, 0, 1 coded CNAs}

\item{loss.col}{colour of losses, defaults to 'blue'}

\item{gain.col}{colour of gains, defaults to 'red'}

\item{neutral.col}{colour of neutral regions, defaults to 'white'}
}
\value{
vector of colours
}
\description{
Convert vector of -1, 0, 1 CNA calls to vector of colours
}
