% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{conf.int.string}
\alias{conf.int.string}
\title{Make confidence interval string.}
\usage{
conf.int.string(confidence.interval, n.decimals = 3)
}
\arguments{
\item{confidence.interval}{vector of length two giving confidence interval.}

\item{n.decimals}{number of decimal places to include in confidence interval.}
}
\value{
formatted.conf.int confidence interval formatted as a string.
}
\description{
Make confidence interval string.
}
