% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{confidence.interval}
\alias{confidence.interval}
\title{confidence.interval}
\usage{
confidence.interval(x, alpha = 0.05)
}
\description{
Get the confidence interval of the mean
 
 @param x vector of numeric values
 @param alpha 1 minus the confidence level
 
 @return vector of length 2, containing lower and upper confidence bound
}
