% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.time.numeric.R
\name{convert.time.numeric}
\alias{convert.time.numeric}
\title{convert.time.numeric}
\usage{
convert.time.numeric(x)
}
\arguments{
\item{x}{Times to be converted}
}
\value{
vector of length \code{x} containing corresponding numeric values
}
\description{
Convert a time in XX:YY format to a numeric value. For example, 12:30 gets converted to 12.5.
}
