% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.similarity.R
\name{cosine.similarity}
\alias{cosine.similarity}
\title{Calculate cosine similarity between two vectors}
\usage{
cosine.similarity(u, v)
}
\arguments{
\item{u}{Numeric vector}

\item{v}{Numeric vector}
}
\value{
calculated cosine similarity between u and v
}
\description{
Calculate cosine similarity between two vectors
}
\examples{
cosine.similarity(1:3, 4:6);

}
\author{
Erle Holgersen <Erle.Holgersen@gmail.com>
}
