% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{factor.to.numeric}
\alias{factor.to.numeric}
\title{Convert a factor to the numeric values in factor levels}
\usage{
factor.to.numeric(x)
}
\arguments{
\item{x}{Factor to be converted}
}
\value{
numeric.values Vector of numeric values corresponding to x
}
\description{
Convert a factor to the numeric values in factor levels
}
