% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{fg.model.ipcw}
\alias{fg.model.ipcw}
\title{IPCW for Fine-Gray model}
\usage{
fg.model.ipcw(t, survival.object, censoring.status = 0)
}
\arguments{
\item{t}{time weights should be calculated at.}

\item{survival.object}{survival object containing survival information for each patient.}

\item{censoring.status}{status level corresponding to censoring, defaults to 0.}
}
\description{
Calculate inverse probability of censoring weighting for Fine and Gray model.
}
\references{
Jason P. Fine and Robert J. Gray (1999). A Proportional Hazards Model for the Subdistribution of a Competing Risk. Journal of the American Statistical Association.
}

