% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.p.value.R
\name{format.p.value}
\alias{format.p.value}
\title{format.p.value}
\usage{
\method{format}{p.value}(p.value, prefix = "P")
}
\arguments{
\item{p.value}{p-value to be formatted}

\item{prefix}{prefix for p-value, defaults to "P"}
}
\value{
language object with p-value expression
}
\description{
Format a p-value for a plot
}
