% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{geometric.mean}
\alias{geometric.mean}
\title{Calculate geometric mean}
\usage{
geometric.mean(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
geometric mean
}
\description{
Calculate geometric mean
}
\author{
Erle Holgersen <erle.holgersen@gmail.com>
}
