% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.colour.palette.R
\name{get.colour.palette}
\alias{get.colour.palette}
\title{get.colour.palette}
\usage{
get.colour.palette(n, palette = c("default", "pastel"))
}
\arguments{
\item{n}{Number of colours to retrieve}

\item{palette}{Name of palette}
}
\value{
Vector of colour HEX codes
}
\description{
Get a set of colours for plotting
}
\author{
Erle Holgersen <Erle.Holgersen@gmail.com>
}
