% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.cat.R
\name{head.cat}
\alias{head.cat}
\title{Print head of an arbitrary number of objects.}
\usage{
\method{head}{cat}(...)
}
\arguments{
\item{...}{objects to print head of}
}
\value{
None
}
\description{
Print head of an arbitrary number of objects.
}
\examples{
a <- 1:100;
b <- 50:100;
head.cat(a, b, letters);

}
