% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer.coded.R
\name{integer.coded}
\alias{integer.coded}
\title{Integer code a vector}
\usage{
integer.coded(x)
}
\arguments{
\item{x}{vector to be converted}
}
\value{
integer-coded vector
}
\description{
Convert a vector to a vector of integers, where each integer corresponds to a unique value in the original vector. Shorthand for as.numeric(factor(x, levels = unique(x)))
}
