% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{interleave}
\alias{interleave}
\title{Interleave two vectors. Any overhang due to mismatching lengths is added onto the end.}
\usage{
interleave(x, y)
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}
}
\value{
Vector of interleaved values.
}
\description{
Interleave two vectors. Any overhang due to mismatching lengths is added onto the end.
}
