% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.error.R
\name{is.error}
\alias{is.error}
\title{is.error}
\usage{
is.error(x)
}
\arguments{
\item{x}{object}
}
\value{
logical indicating if object is an error
}
\description{
Check if an object is an error. Stolen from \url{http://adv-r.had.co.nz/Exceptions-Debugging.html}.
}
