% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{lhs}
\alias{lhs}
\title{Latin hypercube sampling}
\usage{
lhs(s, n)
}
\arguments{
\item{s}{number of dimensions}

\item{n}{number of points in each dimension}
}
\value{
\code{s} by \code{n} dimension matrix of generated points
}
\description{
Latin hypercube sampling
}
\author{
Erle Holgersen <Erle.Holgersen@gmail.com>
}
