% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.tcga.id.R
\name{parse.tcga.id}
\alias{parse.tcga.id}
\title{parse.tcga.id}
\usage{
parse.tcga.id(id, field = c("project", "tss", "participant", "sample", "vial",
  "portion", "analyte", "plate", "center"))
}
\arguments{
\item{id}{TCGA ID (must start with TCGA)}

\item{field}{Name of field to extract}
}
\value{
Extracted component of ID
}
\description{
Extract a specific component from a TCGA ID. See \url{https://wiki.nci.nih.gov/display/TCGA/TCGA+barcode} for details.
}
\author{
Erle Holgesen <Erle.Holgersen@gmail.com>
}
