% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.header.R
\name{read.header}
\alias{read.header}
\title{read.header}
\usage{
read.header(file, sep = "\\t")
}
\arguments{
\item{file}{Path to file}

\item{sep}{Field separator character}
}
\value{
Character vector of header components
}
\description{
Read the header of a file in table format.
}
\author{
Erle Holgersen <Erle.Holgersen@gmail.com>
}
