% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{revd}
\alias{revd}
\title{Generate obervations from extreme value distribution}
\usage{
revd(n, location = 0, scale = 1)
}
\arguments{
\item{n}{number of observations to generate}

\item{location}{location parameter}

\item{scale}{scale parameter}
}
\value{
vector of simulated observations
}
\description{
Generate obervations from extreme value distribution
}
