% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{rnorm.bivariate}
\alias{rnorm.bivariate}
\title{rnorm.bivariate}
\usage{
rnorm.bivariate(n, rho)
}
\arguments{
\item{n}{Number of samples to be generated}

\item{rho}{Correlation between variables}
}
\value{
\code{n} by 2 dimensional matrix of generated numbers
}
\description{
Generate samples from a bivariate normal distribution based on correlation
}
\author{
Erle Holgersen <Erle.Holgersen@gmail.com>
}
