% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific_notation.R
\name{scientific.notation}
\alias{scientific.notation}
\title{scientific.notation}
\usage{
scientific.notation(x)
}
\arguments{
\item{x}{Number to be converted}
}
\value{
a language object
}
\description{
Convert a number to a scientific notation language object
}
\author{
Erle Holgersen <Erle.Holgersen@gmail.com>
}
