% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.colour.palette.R
\name{show.colour.palette}
\alias{show.colour.palette}
\alias{show.color.palette}
\title{show.colour.palette}
\usage{
show.colour.palette(colours)

show.color.palette(colours)
}
\arguments{
\item{colours}{vector of colours}
}
\description{
Show colour palette in a grid
}
\examples{
show.colour.palette( get.colour.palette(7) );
show.colour.palette( colours() );

}
\author{
Erle Holgersen <Erle.Holgersen@gmail.com>
}
