% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{source.directory}
\alias{source.directory}
\title{Source directory}
\usage{
source.directory(directory)
}
\arguments{
\item{directory}{Path to directory to be sourced}
}
\value{
None
}
\description{
Source all R scripts in a directory.
}
\author{
Erle Holgersen <Erle.Holgersen@gmail.com>
}
