% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.cat.R
\name{str.cat}
\alias{str.cat}
\title{Display structure of an arbitrary number of objects.}
\usage{
\method{str}{cat}(...)
}
\arguments{
\item{...}{objects to display structure of}
}
\value{
None
}
\description{
Display structure of an arbitrary number of objects.
}
