% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{truncate.survival}
\alias{truncate.survival}
\title{Truncate survival}
\usage{
\method{truncate}{survival}(t, survival.object, censoring.status = 0)
}
\arguments{
\item{t}{truncation time.}

\item{survival.object}{survival object containing survival information for each patient.}

\item{censoring.status}{status level corresponding to censoring, defaults to 0.}
}
\value{
truncated.survival.object  survival object of same length input, containing survival information as observed at time t.
}
\description{
Truncate survival times to time t.
}
