% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varied_rep.R
\name{varied.rep}
\alias{varied.rep}
\title{rep function extended to support variable number of repeats for each element.}
\usage{
varied.rep(x, each)
}
\arguments{
\item{x}{elements to be repeated.}

\item{each}{vector giving number of times each element should be repeated.}
}
\value{
repeat.vector vector created by repeating each element.
}
\description{
rep function extended to support variable number of repeats for each element.
}
