% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datestamp_filename.R
\name{datestamp.filename}
\alias{datestamp.filename}
\title{Generate date-stamped filename}
\usage{
datestamp.filename(file.name, extension, timestamp = FALSE)
}
\arguments{
\item{file.name}{core part of file name to be date stamped}

\item{extension}{extension}
}
\value{
generated file name
}
\description{
Generate date-stamped filename
}
