% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_plotting.R
\name{km.plot}
\alias{km.plot}
\title{Plot Kaplan-Meier curves.}
\usage{
km.plot(km.formula, dataset, show.risk.table = FALSE, lty = NULL)
}
\arguments{
\item{km.formula}{Formula to be plotted}

\item{dataset}{Dataframe}

\item{show.risk.table}{Boolean indicating whether to show at risk data}

\item{lty}{Line type}
}
\description{
Plot Kaplan-Meier curves.
}
