% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_plotting.R
\name{km.plot}
\alias{km.plot}
\title{Plot Kaplan-Meier curves.}
\usage{
km.plot(km.formula, dataset, show.risk.table = FALSE)
}
\arguments{
\item{km.formula}{Formula to be plotted.}

\item{dataset}{Data.}

\item{show.risk.table}{Boolean.}
}
\description{
Plot Kaplan-Meier curves.
}
