% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripboxplot.R
\name{stripboxplot}
\alias{stripboxplot}
\title{stripboxplot}
\usage{
stripboxplot(plot.formula, dataset = NULL, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{plot.formula}{Formula for plotting}

\item{dataset}{Dataset}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{...}{additional parameters to be passed to \code{stripchart}}
}
\description{
Make a boxplot with underlying stripplot
}
