% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripboxplot.R
\name{stripboxplot}
\alias{stripboxplot}
\title{stripboxplot}
\usage{
stripboxplot(plot.formula, dataset = NULL, xlab = NULL, ylab = NULL,
  frame = FALSE, points.col = "lightgrey", ...)
}
\arguments{
\item{plot.formula}{Formula for plotting}

\item{dataset}{Dataset}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{frame}{logical indicating whether to plot full frame}

\item{points.col}{Colour of background points}

\item{...}{Additional parameters to be passed to \code{stripchart}}
}
\description{
Make a boxplot with underlying stripplot
}
\examples{
test.data <- data.frame(
 x = sample(letters[1:3], 50, replace = TRUE),
 y = rnorm(50)
 );

stripboxplot(y ~ x, test.data);

}
\author{
Erle Holgersen <Erle.Holgersen@gmail.com>
}
